requirements.txt
scipy==1.3.1
numpy==1.17.3
pysftp==0.2.9
SQLAlchemy==1.3.10
pandas==0.25.2
workdays==1.4
psycopg2_binary==2.8.4
py_vollib==1.0.1
psycopg2==2.8.6
python_dateutil==2.8.1
